Create a new kim create_box command
Create a new `kim create_box` command with a syntax of,
kim create_box region-ID keyword value ...`, where
`region-ID` is the ID of region to use as simulation domain, and
`keyword` is one of the `extra/bond/per/atom` or
`extra/angle/per/atom` or `extra/dihedral/per/atom` or
`extra/improper/per/atom` or `extra/special/per/atom`.

This command wraps the standard `create_box` command and passes to
it the region-ID argument and the maximum values stored by `kim init`
for all types.

If this is called for a non-bonded SM, it creates an error.

It also sets all masses to a default of 1.0, and it is the user
responsibility to provide the correct masses for used types.