Add a new 'kim create_box' command
Add a new 'kim create_box region-id' command. If this is called for
a non-bonded SM (i.e. flag kim_bonded_ff is not true), generate an
error. For the bonded SMs it wraps the standard 'create_box' command
and pass to it the region-id argument and the maximum values stored
by 'kim init' for all types.